+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * fix issue when changing search indices - the search query was wrongly reset
 * adds buttons to collapse/expand json, fixes [#291](https://github.com/cars10/elasticvue/issues/291)
 * added keyboard shortcuts to search for mac, thanks @prashantsingh3553
+* show number of shards on nodes overview for elasticsearch >= 8, fixes [#287](https://github.com/cars10/elasticvue/issues/287)
 
 ## 1.4.0